+++ tests/integration_sync/devnet/test_srpc_gets.py
@@ -17,11 +17,9 @@ from pysui.sui.sui_types.scalars import ObjectID, SuiString, SuiInteger
 from pysui.sui.sui_types.address import SuiAddress
 from pysui.sui.sui_txresults import (
     ObjectNotExist,
-    MoveDataDescriptor,
     ObjectVersionNotFound,
     ObjectVersionTooHigh,
     SuiData,
-    SuiGasDescriptor,
     SuiGas,
 )
 from pysui.sui.sui_builders.get_builders import GetCoins, GetCommittee, GetObject, GetPastObject, GetCoinTypeBalance
@@ -44,24 +42,6 @@ def get_gas(client: SuiClient, for_address: SuiAddress = None) -> list[SuiGas]:
     # result: SuiRpcResult = client.get_objects_for(ident_list)
     # assert result.is_ok()
     return result.result_data.data
-
-
-def get_data(client: SuiClient, for_address: SuiAddress = None) -> list[SuiData]:
-    """get_data Fetch all data objects for address.
-
-    :param client: Synchronous http client
-    :type client: SuiClient
-    :param for_address: Address to get objects for, defaults to None and uses active-address
-    :type for_address: SuiAddress, optional
-    :return: List of data objects owned by address, maybe empty
-    :rtype: list[SuiData]
-    """
-    result: SuiRpcResult = client.get_address_object_descriptors(MoveDataDescriptor, for_address)
-    assert result.is_ok()
-    ident_list = [desc.identifier for desc in result.result_data]
-    result: SuiRpcResult = client.get_objects_for(ident_list)
-    assert result.is_ok()
-    return result.result_data
 
 
 def test_get_gas_activeaddress_pass(sui_client: SuiClient):
@@ -135,25 +115,6 @@ def test_get_gas_anyaddress_pass(sui_client: SuiClient):
             gas_balances = [gas.balance for gas in gas_objects]
             total_balance = sum(gas_balances)
             assert total_balance > 0
-
-
-def test_get_object_pass(sui_client: SuiClient):
-    """test_get_object_pass Validate succesful fetch for details of gas object.
-
-    :param sui_client: Synchronous http client
-    :type sui_client: SuiClient
-    """
-    gas_objects = get_gas(sui_client)
-    assert gas_objects
-    # Convenience method on client
-    gas_object = sui_client.get_object(gas_objects[0].identifier)
-    assert gas_object
-    assert gas_object.is_ok()
-    # Builder approach
-    builder = GetObject(gas_objects[0].identifier)
-    gas_object2 = sui_client.execute(builder)
-    assert gas_object2.is_ok()
-    assert gas_object.result_data == gas_object2.result_data
 
 
 def test_get_object_fail(sui_client: SuiClient):